maple: audio: enable aaudio and HiFi routes for better playback
Complementing my "prepare" commit in the common dt, we enable the
AAudio and HiFi routes we added there through the mixer.

AAudio is a new playback method introduced in Oreo for low latency with
the ability to write directly to ALSA buffers. Explained here
https://web.archive.org/web/20190531163856/https://source.android.com/devices/audio/aaudio.
HiFi playback routes are from stock to allow improved playback with
apps that do not dedicate exclusive access to the audio platform.

Signed-off-by: Laster K. (lazerl0rd) <officiallazerl0rd@gmail.com>